F #2940: GOCA - fix id are signed integers

+++ src/oca/go/src/goca/acl.go
@@ -25,7 +25,7 @@ type ACLsController entitiesController
 
 // ACLPool represents an OpenNebula ACL list pool
 type ACLPool struct {
-	ID       uint   `xml:"ID"`
+	ID       int    `xml:"ID"`
 	User     int    `xml:"USER"`
 	Resource int    `xml:"RESOURCE"`
 	Rights   int    `xml:"RIGHTS"`
@@ -100,17 +100,17 @@ func (ac *ACLsController) Info() (*ACLPool, error) {
 // * user: User component of the new rule. A string containing a hex number.
 // * resource: Resource component of the new rule. A string containing a hex number.
 // * rights: Rights component of the new rule. A string containing a hex number.
-func (ac *ACLsController) CreateRule(user AclUsers, resource AclResources, rights AclRights) (uint, error) {
+func (ac *ACLsController) CreateRule(user AclUsers, resource AclResources, rights AclRights) (int, error) {
 	response, err := ac.c.Client.Call("one.acl.addrule", user, resource, rights)
 	if err != nil {
 		return 0, err
 	}
 
-	return uint(response.BodyInt()), nil
+	return response.BodyInt(), nil
 }
 
 // DeleteRule deletes an ACL rule.
-func (ac *ACLsController) DeleteRule(aclID uint) error {
-	_, err := ac.c.Client.Call("one.acl.delrule", int(aclID))
+func (ac *ACLsController) DeleteRule(aclID int) error {
+	_, err := ac.c.Client.Call("one.acl.delrule", aclID)
 	return err
 }
